+++ modules/core/import/modules/csv/tests/src/Kernel/CsvImportMigrationGroupTest.php
@@ -17,8 +17,10 @@ class CsvImportMigrationGroupTest extends CsvImportTestBase {
   protected static $modules = [
     'farm_animal_type',
     'farm_equipment',
+    'farm_equipment_type',
     'farm_harvest',
     'farm_id_tag',
+    'taxonomy',
   ];
 
   /**
@@ -27,6 +29,7 @@ class CsvImportMigrationGroupTest extends CsvImportTestBase {
   public function setUp(): void {
     parent::setUp();
     $this->installConfig(['farm_animal_type']);
+    $this->installConfig(['farm_equipment_type']);
     $this->installConfig(['farm_equipment']);
     $this->installConfig(['farm_harvest']);
     $this->installConfig(['farm_import_csv_test']);