+++ modules/quick/group/tests/src/Kernel/QuickGroupTest.php
@@ -27,9 +27,11 @@ class QuickGroupTest extends QuickFormTestBase {
    */
   protected static $modules = [
     'farm_equipment',
+    'farm_equipment_type',
     'farm_group',
     'farm_observation',
     'farm_quick_group',
+    'taxonomy',
   ];
 
   /**
@@ -38,6 +40,7 @@ class QuickGroupTest extends QuickFormTestBase {
   protected function setUp(): void {
     parent::setUp();
     $this->installConfig([
+      'farm_equipment_type',
       'farm_equipment',
       'farm_group',
       'farm_observation',