+++ modules/quick/inventory/tests/src/Kernel/QuickInventoryTest.php
@@ -29,6 +29,7 @@ class QuickInventoryTest extends QuickFormTestBase {
   protected static $modules = [
     'farm_activity',
     'farm_equipment',
+    'farm_equipment_type',
     'farm_inventory',
     'farm_observation',
     'farm_quantity_standard',
@@ -43,6 +44,7 @@ class QuickInventoryTest extends QuickFormTestBase {
     parent::setUp();
     $this->installConfig([
       'farm_activity',
+      'farm_equipment_type',
       'farm_equipment',
       'farm_observation',
       'farm_quantity_standard',